refactor: io_thread -> root/iothread.c
