+++ Makefile.am
@@ -70,6 +70,7 @@ watchman_SOURCES = \
 	clientmode.c \
 	main.c       \
 	root.c       \
+	root/iothread.c       \
 	root/lock.c       \
 	root/symlink.c       \
 	state.c      \
